Project.toml and project refactoring
