+++ .gitignore
@@ -59,3 +59,10 @@ target/
 
 # VS Code
 .vscode/
+
+# PDM
+.pdm-python
+.venv
+
+# Misc
+.mypy_cache
