+++ README.md
@@ -1,0 +1,5 @@
+# bson
+
+Forked from: https://github.com/py-bson/bson.
+
+Removed external dependencies, Python 2 compatibility, ObjectId implementation, and network utility.
