+++ src/zaber_bson/__init__.py
@@ -44,25 +44,3 @@ def loads(data):
         Given a BSON string, outputs a dict.
     """
     return decode_document(data, 0)[1]
-
-
-def patch_socket():
-    """
-        Patches the Python socket class such that sockets can send and receive BSON
-        objects atomically.
-
-        This adds the following functions to socket:
-
-        recvbytes(bytes_needed, sock_buf = None) - reads bytes_needed bytes
-        atomically. Returns None if socket closed.
-
-        recvobj() - reads a BSON document from the socket atomically and returns
-        the deserialized dictionary. Returns None if socket closed.
-
-        sendobj(obj) - sends a BSON document to the socket atomically.
-    """
-    from socket import socket
-    from .network import recvbytes, recvobj, sendobj
-    socket.recvbytes = recvbytes
-    socket.recvobj = recvobj
-    socket.sendobj = sendobj
