+++ test/test_datetime.py
@@ -2,7 +2,7 @@
 from datetime import datetime, timezone
 from unittest import TestCase
 
-from bson import dumps, loads
+from zaber_bson import dumps, loads
 
 
 class TestDateTime(TestCase):
