+++ test/test_int.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 from unittest import TestCase
 
-from bson import dumps, loads
+from zaber_bson import dumps, loads
 
 
 class TestInt(TestCase):
