+++ test/test_int_as_key.py
@@ -1,6 +1,6 @@
 from unittest import TestCase
 
-from bson import dumps, loads
+from zaber_bson import dumps, loads
 import json
 
 class TestIntAsKey(TestCase):
@@ -9,12 +9,10 @@ class TestIntAsKey(TestCase):
             10: 'pass?',
             10.2: 'pass?'
         }
-    
+
     def test_int_as_key(self):
         dump = dumps(self.test_dict)
         decoded = loads(dump)
         dumpByJson = json.dumps(self.test_dict)
         decodedByJson = json.loads(dumpByJson)
         self.assertEqual(decoded, decodedByJson)
-
-
